Update CRT and NTT

+++ Math/NumberTheory/ChineseRemainder.h
@@ -2,13 +2,19 @@
 // - a[i] * x = b[i] MOD m[i] (mi don't need to be co-prime)
 // Tested:
 // - https://open.kattis.com/problems/generalchineseremainder
-bool linearCongruences(const vector<ll> &a, const vector<ll> &b,
-        const vector<ll> &m, ll &x, ll &M) {
-    ll n = a.size();
+// - https://oj.vnoi.info/problem/icpc21_mt_d
+bool linearCongruences(
+        const vector<long long> &a,
+        const vector<long long> &b,
+        const vector<long long> &m,
+        long long &x,
+        long long &M
+) {
+    long long n = a.size();
     x = 0; M = 1;
     REP(i, n) {
-        ll a_ = a[i] * M, b_ = b[i] - a[i] * x, m_ = m[i];
-        ll y, t, g = extgcd(a_, m_, y, t);
+        long long a_ = a[i] * M, b_ = b[i] - a[i] * x, m_ = m[i];
+        long long y, t, g = extgcd(a_, m_, y, t);
         if (b_ % g) return false;
         b_ /= g; m_ /= g;
         x += M * (y * b_ % m_);
